fix(workflow,skill): phase 8 h wave 2 — cross-user install isolation, group enforcement, durable cancel, entry_point safety, caps + tests
H1: owner-scope BOTH the on-disk path (<data>/{skills,workflows}/<owner-or-system>/<name>/<ver>/)
and the unique constraint (migration 95: per-owner partial unique indexes replacing the global
UNIQUE(name,version)); re-install now overwrites cleanly (delete prior owner row + hub_entities,
re-insert) instead of 23505.
H6: dev import namespaces the slug per user (local.dev.<owner>/<slug>) + owner-scopes the path +
filters the pre-delete by owner.
M6: system skill/workflow installs wrap insert + group rows + hub_entities track in one TX
(install_system_{skill,workflow}_tx); extracted dir cleaned on any post-extract error; fixed the
misleading "one TX" doc.
SEC-2: validate manifest entry_point at install time (reject .., absolute, separators) via
bundle::validate_entry_point before any join/persist.
H2: workflow list_for_user now applies the group-restriction predicate (mirrors skill); added
repository::user_can_access + wired it into workflow GET/RUN so a group-restricted system
workflow 404s for non-members (also covers workflow_mcp tool-list/call via list_for_user).
H3: guarded mark_status (status NOT IN terminal) + mark_running (status='pending') so a late
step-complete can't clobber cancelled and a fast cancel isn't resurrected to running.
M3: artifact collector enforces the per-run cap PRE-WRITE (checks metadata size before reading
the body); runner fails the run instead of swallowing the cap error.
M4: total_tokens INTEGER -> BIGINT (migration 95 + WorkflowRun.total_tokens i64 + checked
i64 cast in persist_step_meta).
M2: glob_match treats / as a boundary (single * no longer crosses /, ** does); match_decl
prefers exact path: decls over glob: in a two-pass match.
M1: removed the unused log_io SPILL_THRESHOLD_BYTES + corrected the always-spill doc.
L1: char-safe truncation in resolve_outputs (was a byte slice &rendered[..500] that panics
mid-codepoint).
L2: parenthesized the find_by_name_version (name,version) predicate (skill + workflow).
L4: workflow delete now removes the disk dir AFTER the DB delete (matches skill).
Consistency: unified the validate-error field name to `location` (skill was `path`); fixed the
migration 95 comment that falsely claimed a linkme slice in permissions.rs. SKIPPED the
skill/handlers.rs file-split into handlers/{mod,system,dev}.rs — purely cosmetic, high churn
across routes.rs + hub re-exports; prioritized correctness/security per the brief.

Tests: cross-user list isolation (skill + workflow), group-restricted system workflow
invisible/unrunnable to a non-member (H2), cancel-then-complete durability (H3), entry_point
traversal rejected at install (SEC-2), bundle bomb cumulative + file-count caps (bundle unit),
dispatch shell_escape_single + retry_backoff + on_error classify branches; fixed the
false-coverage run_with_mocks_on_published_workflow_is_rejected to install a real PUBLISHED
(non-dev) workflow and assert the 403.

+++ src-app/server/src/modules/workflow/models.rs
@@ -124,7 +124,7 @@ pub struct WorkflowRun {
     pub status: String,
     pub current_step: Option<String>,
     pub error_message: Option<String>,
-    pub total_tokens: i32,
+    pub total_tokens: i64, // M4: BIGINT column — a long run can exceed i32 range
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
 }
